refactor(nominations): deduplicate evaluate_org wrapper into claim module
